[wasm] Fix MaxPool for 1x1 kernels
Create identity_pool_test with a function to test similar ops against 1x1
kernels and use it to test MaxPool. Fix MaxPool in the wasm backend.
